requirements.txt
selenium>=4.0.0
requests>=2.25.0
aiohttp>=3.8.0
psutil>=5.8.0
watchdog>=2.1.0
jsonschema>=4.0.0
webdriver_manager>=3.5.0
fake_useragent>=0.1.11
python-daemon>=2.3.0